feat: use go modules to manage dependencies
This introduces use of go modules to manage our dependency tree, ensuring
we all use the exact same version of our direct and indirect dependencies
when building.

Although as of writing this, go modules is in its early experimentation
stages, I haven't found any critical errors reported yet and the core go
team has promised to be backwards compatible going forward. With that in
mind in addition to knowing go modules will be the way dependencies are
handled by the go community in the future, especially after being marked
as stable in the next release or two, it sounds quite natural to give it
a shot.

Refs https://github.com/golang/go/wiki/Modules