fix(ci): introduce explicit dependency install instructions to Travis
This is primarily done to ensure all dependencies has been installed
via go modules *before* any of the checks (vetting and formatting) are
executed.

Our `go fmt` check strictly requires *nothing* to be written to stdout
as long as all files are formatted properly. If go modules hasn't had
the chance to install dependencies before `go fmt` is run, the go modules
install procedure will kick in which causes quite a lot of verbose
info to be written to stdout.

Although go modules ensuring things are already installed before running
most `go` CLI commands, it's more reliable for us to just be explicit
about it and ensure that install procedure gets run as a separate thing
to any of other commands we execute.